requirement.txt
absl-py==0.12.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cryptography @ file:///C:/ci/cryptography_1615532620702/work
cycler==0.10.0
defusedxml==0.7.1
df2gspread==1.0.4
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.14
google-api-python-client==1.6.7
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gspread==3.7.0
gspread-dataframe==3.2.1
httplib2==0.19.0
idna==2.10
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==3.7.3
ipywidgets==7.6.3
jedi==0.17.0
Jinja2==2.11.3
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema==3.2.0
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver @ file:///C:/ci/kiwisolver_1612282618948/work
Markdown==3.3.4
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mysql-connector-python==8.0.18
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
prometheus-client==0.9.0
protobuf==3.15.6
pyaml==20.4.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.6.1
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
PyMySQL @ file:///C:/ci/pymysql_1610464946597/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent==0.17.3
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1612458529756/work
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn @ file:///C:/ci/scikit-learn_1614446716349/work
scipy @ file:///C:/ci/scipy_1614023124510/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
Send2Trash==1.5.0
six @ file:///C:/ci/six_1605205426665/work
smmap==3.0.5
SQLAlchemy @ file:///C:/ci/sqlalchemy_1612467804654/work
streamlit==0.79.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
toolz==0.11.1
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
validators==0.18.2
watchdog==2.0.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xgboost==0.90
zipp==3.4.1
